refactor(lib/page-routing): rename lib/effector, add StartParams type
