+++ src/lib/page-routing.ts
@@ -2,15 +2,18 @@ import * as React from 'react';
 import { Event } from 'effector';
 
 const START = `☄️/start-event`;
-type Params = Record<string, string>;
+export type StartParams = {
+  params: Record<string, string>;
+  query: Record<string, string>;
+};
 
-export function getStart<T>(component: T): undefined | Event<Params> {
+export function getStart<T>(component: T): undefined | Event<StartParams> {
   return component[START];
 }
 
 export function assignStart(
   component: React.Component | React.FC,
-  event: Event<Params>,
+  event: Event<StartParams>,
 ) {
   component[START] = event;
 }
